// sdk/typescript-sdk/src/models/MakeEnvResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plato API
 * API for Plato platform - SDK routes only
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MakeEnvResponse
 */
export interface MakeEnvResponse {
    /**
     * The job ID for the created environment
     * @type {string}
     * @memberof MakeEnvResponse
     */
    jobId: string;
    /**
     * Current status of the environment
     * @type {string}
     * @memberof MakeEnvResponse
     */
    status?: string;
}

/**
 * Check if a given object implements the MakeEnvResponse interface.
 */
export function instanceOfMakeEnvResponse(value: object): value is MakeEnvResponse {
    if (!('jobId' in value) || value['jobId'] === undefined) return false;
    return true;
}

export function MakeEnvResponseFromJSON(json: any): MakeEnvResponse {
    return MakeEnvResponseFromJSONTyped(json, false);
}

export function MakeEnvResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MakeEnvResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'jobId': json['job_id'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function MakeEnvResponseToJSON(json: any): MakeEnvResponse {
    return MakeEnvResponseToJSONTyped(json, false);
}

export function MakeEnvResponseToJSONTyped(value?: MakeEnvResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'job_id': value['jobId'],
        'status': value['status'],
    };
}