// sdk/typescript-sdk/src/models/CreateVMResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plato API
 * API for Plato platform - SDK routes only
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateVMResponse
 */
export interface CreateVMResponse {
    /**
     * Status
     * @type {string}
     * @memberof CreateVMResponse
     */
    status: string;
    /**
     * 
     * @type {string}
     * @memberof CreateVMResponse
     */
    timestamp?: string | null;
    /**
     * Correlation ID for tracking
     * @type {string}
     * @memberof CreateVMResponse
     */
    correlationId: string;
    /**
     * Public URL for the VM
     * @type {string}
     * @memberof CreateVMResponse
     */
    url: string;
    /**
     * Public job ID
     * @type {string}
     * @memberof CreateVMResponse
     */
    jobPublicId: string;
    /**
     * Job group ID
     * @type {string}
     * @memberof CreateVMResponse
     */
    jobGroupId: string;
}

/**
 * Check if a given object implements the CreateVMResponse interface.
 */
export function instanceOfCreateVMResponse(value: object): value is CreateVMResponse {
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('correlationId' in value) || value['correlationId'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    if (!('jobPublicId' in value) || value['jobPublicId'] === undefined) return false;
    if (!('jobGroupId' in value) || value['jobGroupId'] === undefined) return false;
    return true;
}

export function CreateVMResponseFromJSON(json: any): CreateVMResponse {
    return CreateVMResponseFromJSONTyped(json, false);
}

export function CreateVMResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateVMResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'],
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
        'correlationId': json['correlation_id'],
        'url': json['url'],
        'jobPublicId': json['job_public_id'],
        'jobGroupId': json['job_group_id'],
    };
}

export function CreateVMResponseToJSON(json: any): CreateVMResponse {
    return CreateVMResponseToJSONTyped(json, false);
}

export function CreateVMResponseToJSONTyped(value?: CreateVMResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'timestamp': value['timestamp'],
        'correlation_id': value['correlationId'],
        'url': value['url'],
        'job_public_id': value['jobPublicId'],
        'job_group_id': value['jobGroupId'],
    };
}